+++ include/reconstructor/calc_projs_mt/apply_chunk_to_proj.h
@@ -6,8 +6,8 @@
 /* 
  * 
  */
-void apply_chunk_to_proj(double* dev_chunk, unsigned int dim_chunk,
-  double* dev_chunk_origin, double* dev_projs, double* dev_r_hats,
+void apply_chunk_to_proj(float* dev_chunk, unsigned int dim_chunk,
+  float* dev_chunk_origin, float* dev_projs, float* dev_r_hats,
   unsigned int num_projs, unsigned int dim_proj_x, unsigned int dim_proj_y,
   unsigned int lim_proj_z);
