+++ src/data/data_ops.c
@@ -2,7 +2,7 @@
 
 #include <math.h>
 
-void euler_rot(double* v, double* a, double* v_f){
+void euler_rot(float* v, float* a, float* v_f){
   v_f[0] = 
     v[0]*cos(a[0])*cos(a[1]) + 
     v[2]*(cos(a[0])*cos(a[2])*sin(a[1]) + sin(a[0])*sin(a[2])) + 
@@ -17,7 +17,7 @@ void euler_rot(double* v, double* a, double* v_f){
     v[1]*cos(a[1])*sin(a[2]);
 }
 
-void euler_rot_rev(double* v, double* a, double* v_f){
+void euler_rot_rev(float* v, float* a, float* v_f){
   v_f[0] = 
     v[0]*cos(a[0])*cos(a[1]) +
     v[1]*cos(a[1])*sin(a[0]) - 
@@ -33,7 +33,7 @@ void euler_rot_rev(double* v, double* a, double* v_f){
 }
 
 void deg_to_rad(Data_2d* a){
-  double f = PI/180.;
+  float f = PI/180.;
   for(unsigned int i=0; i<(a->dim)[0]; i++){
     for(unsigned int j=0; j<(a->dim)[1]; j++){
       (a->data)[i][j] = (a->data)[i][j]*f;
@@ -42,7 +42,7 @@ void deg_to_rad(Data_2d* a){
 }
 
 void rad_to_deg(Data_2d* a){
-  double f = 180./PI;
+  float f = 180./PI;
   for(unsigned int i=0; i<(a->dim)[0]; i++){
     for(unsigned int j=0; j<(a->dim)[1]; j++){
       (a->data)[i][j] = (a->data)[i][j]*f;
