+++ src/reconstructor/apply_projs_diff_mt/apply_proj_to_chunk.cu
@@ -1,5 +1,5 @@
 extern "C"{
-  #include <reconstructor/apply_projs_diff_gpu/apply_proj_to_chunk.h>
+  #include <reconstructor/apply_projs_diff_mt/apply_proj_to_chunk.h>
 
   #include <data/data_ops.h>
   #include <math.h>
@@ -12,13 +12,13 @@ extern "C"{
 __device__ int in_bounds_gpu(int i, int Ni){
   return (i>=0 && i<Ni);
 }
-__device__ double proj_interp_val_gpu(double* projs_diff, unsigned int p_idx,
-  unsigned int Ni, unsigned int Nj, double x, double y){
-  double x_0 = x-floor(x);
-  double y_0 = y-floor(y);
+__device__ float proj_interp_val_gpu(float* projs_diff, unsigned int p_idx,
+  unsigned int Ni, unsigned int Nj, float x, float y){
+  float x_0 = x-floor(x);
+  float y_0 = y-floor(y);
   int i = floor(x);
   int j = floor(y);
-  double q11, q12, q21, q22;
+  float q11, q12, q21, q22;
   if(in_bounds_gpu(i,Ni) && in_bounds_gpu(j,Nj)){
     q11 = projs_diff[p_idx*Nj*Ni+j*Ni+i]; } else{ q11 = 0.;
   }
@@ -31,15 +31,15 @@ __device__ double proj_interp_val_gpu(double* projs_diff, unsigned int p_idx,
   if(in_bounds_gpu(i+1,Ni) && in_bounds_gpu(j+1,Nj)){
     q22 = projs_diff[p_idx*Nj*Ni+(j+1)*Ni+(i+1)]; } else{ q22 = 0.;
   }
-  double r1 = (1-x_0)*q11+x_0*q21;
-  double r2 = (1-x_0)*q12+x_0*q22;
+  float r1 = (1-x_0)*q11+x_0*q21;
+  float r2 = (1-x_0)*q12+x_0*q22;
   return (1-y_0)*r1+y_0*r2;
 }
 
-__global__ void update_chunk_val(double* projs_diff, unsigned int n_proj,
-  unsigned int n_proj_tot, unsigned int pdx, unsigned int pdy, double* chunk,
+__global__ void update_chunk_val(float* projs_diff, unsigned int n_proj,
+  unsigned int n_proj_tot, unsigned int pdx, unsigned int pdy, float* chunk,
   unsigned int dim_chunk, unsigned int x0, unsigned int y0, unsigned int z0,
-  unsigned int vdx, unsigned int vdy, unsigned int vdz, double* r_hats){
+  unsigned int vdx, unsigned int vdy, unsigned int vdz, float* r_hats){
 
   unsigned int tid_shift = 0;                             // device_idx*dim_chunk*dim_chunk*dim_chunk/num_devices;
   unsigned int tid_max   = dim_chunk*dim_chunk*dim_chunk; // (device_idx+1)*dim_chunk*dim_chunk*dim_chunk/num_devices;
@@ -47,16 +47,16 @@ __global__ void update_chunk_val(double* projs_diff, unsigned int n_proj,
 
   while(tid < tid_max){
     int i, j, k, x_center, y_center, z_center, p_x_center, p_y_center;
-    double x, y, z, rx, ry, v;
+    float x, y, z, rx, ry, v;
     k = tid/(dim_chunk*dim_chunk);
     j = (tid-k*dim_chunk*dim_chunk)/dim_chunk;
     i = tid-k*dim_chunk*dim_chunk-j*dim_chunk;
     z_center = vdz/2;
     y_center = vdy/2;
     x_center = vdx/2;
-    z = (double)(k+(int)z0-z_center);
-    y = (double)(j+(int)y0-y_center);
-    x = (double)(i+(int)x0-x_center);
+    z = (float)(k+(int)z0-z_center);
+    y = (float)(j+(int)y0-y_center);
+    x = (float)(i+(int)x0-x_center);
 
     p_x_center = pdx/2;
     p_y_center = pdy/2;
@@ -78,10 +78,10 @@ __global__ void update_chunk_val(double* projs_diff, unsigned int n_proj,
 
 }
 
-void apply_proj_to_chunk(double* dev_projs_diff, unsigned int n_proj,
-  unsigned int n_proj_tot, unsigned int pdx, unsigned int pdy, double* dev_chunk,
+void apply_proj_to_chunk(float* dev_projs_diff, unsigned int n_proj,
+  unsigned int n_proj_tot, unsigned int pdx, unsigned int pdy, float* dev_chunk,
   unsigned int dim_chunk, unsigned int x0, unsigned int y0, unsigned int z0,
-  unsigned int vdx, unsigned int vdy, unsigned int vdz, double* dev_r_hats){
+  unsigned int vdx, unsigned int vdy, unsigned int vdz, float* dev_r_hats){
 
   update_chunk_val<<<BLOCK_COUNT,THREAD_COUNT>>>(
     dev_projs_diff, n_proj, n_proj_tot, pdx, pdy, dev_chunk, dim_chunk,
